+++ .gitignore
@@ -31,3 +31,5 @@ local.properties
 .gradle
 /app/google-services.json
 
+# MacOS
+*.DS_Store
