+++ app/build.gradle.kts
@@ -132,6 +132,7 @@ dependencies {
     implementation(platform(libs.compose.bom))
 
     implementation(libs.firebase.auth.ktx)
+    implementation(libs.androidx.navigation.compose)
 
     testImplementation(libs.junit)
     globalTestImplementation(libs.androidx.junit)
