+++ app/src/main/java/com/android/feedme/MainActivity.kt
@@ -12,10 +12,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
-import com.android.feedme.profile.ProfileScreen
+import androidx.navigation.compose.rememberNavController
 import com.android.feedme.resources.C
 import com.android.feedme.ui.auth.LoginScreen
 import com.android.feedme.ui.camera.CameraScreen
+import com.android.feedme.ui.navigation.NavigationActions
+import com.android.feedme.ui.profile.ProfileScreen
 import com.android.feedme.ui.theme.feedmeAppTheme
 
 class MainActivity : ComponentActivity() {
@@ -40,11 +42,13 @@ class MainActivity : ComponentActivity() {
         Surface(
             modifier = Modifier.fillMaxSize().semantics { testTag = C.Tag.main_screen_container },
             color = MaterialTheme.colorScheme.background) {
+              val navController = rememberNavController()
+              val navigationActions = NavigationActions(navController)
               when (currentScreen) {
                 CurrentScreen.LOGIN -> LoginScreen()
                 CurrentScreen.CAMERA -> CameraScreen()
                 CurrentScreen.LANDING -> TODO()
-                CurrentScreen.PROFILE -> ProfileScreen()
+                CurrentScreen.PROFILE -> ProfileScreen(navigationActions)
               }
             }
       }
