+++ gradle/libs.versions.toml
@@ -32,6 +32,7 @@ firebaseFirestoreKtx = "24.11.0"
 gms = "4.4.1"
 navigationRuntimeKtx = "2.7.7"
 camerax = "1.3.0-rc01"
+navigationCompose = "2.7.7"
 
 
 [libraries]
@@ -83,6 +84,7 @@ camera-lifecycle = { group = "androidx.camera", name = "camera-lifecycle", versi
 camera-video = { group = "androidx.camera", name = "camera-video", version.ref = "camerax" }
 camera-view = { group = "androidx.camera", name = "camera-view", version.ref = "camerax" }
 camera-extensions = { group = "androidx.camera", name = "camera-extensions", version.ref = "camerax" }
+androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "navigationCompose" }
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
